Finish jetty example

+++ examples/jetty/src/main/scala/com/acs/scala/server/mad/examples/jetty/Person.scala
@@ -4,5 +4,5 @@ case class Person
 (
   id: Long,
   name: String,
-  age: Int
+  age: Long
 )
